// src/api/model/currentAccount.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { SectorEnum } from './sectorEnum';
import type { CurrentContactInformation } from './currentContactInformation';
import type { BankAccount } from './bankAccount';
import type { CurrentAccountBankAccount } from './currentAccountBankAccount';

export interface CurrentAccount {
  /** @nullable */
  id?: number | null;
  code: string;
  title: string;
  active: boolean;
  sector: SectorEnum;
  contactInformation: CurrentContactInformation;
  bankAccount: BankAccount;
  currentAccountBankAccounts: CurrentAccountBankAccount[];
}